Remove generated files from the repository (mythdvd/mtd stuff), and set the svn:ignore props for them. x86-64 compile fix for mythgallery. The configure script now checks for non-optional dependencies, and disables a plugin if not everything is found.
git-svn-id: http://svn.mythtv.org/svn/trunk@6743 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythgallery/mythgallery/qtiffio.cpp
@@ -74,7 +74,7 @@ static toff_t tiff_size(thandle_t handle) {
 /*! \warning always returns MAP_FAILED.
 */
 static int tiff_mmap(thandle_t, tdata_t* ,toff_t* ) {
-  return (int) MAP_FAILED;
+  return (long) MAP_FAILED;
 }
 
 //! QIODevice / libtiff write wrapper
